初始化项目

+++ ruoyi-admin/src/main/java/com/ruoyi/im/http/entity/ImChatGroup.java
@@ -41,7 +41,7 @@ public class ImChatGroup implements Serializable {
      * 群主用户ID
      */
     @TableField("master_user_id")
-    private Integer masterUserId;
+    private Long masterUserId;
 
     /**
      * 类型[1:聊天群，2:话题]
