+++ ruoyi-admin/src/main/java/com/ruoyi/im/http/vo/GroupMessageVO.java
@@ -19,7 +19,7 @@ public class GroupMessageVO {
     /**
      * 发送ID
      */
-    private Integer fromId;
+    private Long fromId;
 
     /**
      * 接收ID
